+++ lib/models/enum/payment_channel.dart
@@ -1,0 +1,20 @@
+enum PaymentChannel {
+  online,
+  inStore,
+  other,
+}
+
+extension PaymentChannelExtension on PaymentChannel {
+  String get str {
+    switch (this) {
+      case PaymentChannel.online:
+        return 'Online';
+      case PaymentChannel.inStore:
+        return 'In Store';
+      case PaymentChannel.other:
+        return 'Other';
+      default:
+        return 'Other';
+    }
+  }
+}