+++ lib/models/enum/transaction_item_type.dart
@@ -1,0 +1,7 @@
+enum TransactionItemType {
+  item,
+  tax,
+  subtotal,
+  tip,
+  others,
+}