+++ lib/theme/theme_colors.dart
@@ -1,0 +1,47 @@
+import 'dart:ui';
+
+class ThemeColors {
+  static const primary900 = Color(0xff004c31);
+  static const primary800 = Color(0xff00674c);
+  static const primary700 = Color(0xff007759);
+  static const primary600 = Color(0xff008768);
+  static const primary500 = Color(0xff009474);
+  static const primary400 = Color(0xff27a487);
+  static const primary300 = Color(0xff4eb59b);
+  static const primary200 = Color(0xff80cab7);
+  static const primary100 = Color(0xffb1dfd3);
+  static const primary050 = Color(0xffdff2ee);
+
+  static const complementary900 = Color(0xff940020);
+  static const complementary800 = Color(0xffa30d2a);
+  static const complementary700 = Color(0xffb01531);
+  static const complementary600 = Color(0xffc21f37);
+  static const complementary500 = Color(0xffd02739);
+  static const complementary400 = Color(0xffce4050);
+  static const complementary300 = Color(0xffc7646f);
+  static const complementary200 = Color(0xffd98e96);
+  static const complementary100 = Color(0xfff0c6cf);
+  static const complementary050 = Color(0xfff9e8ec);
+
+  static const analogousPrimary900 = Color(0xff006404);
+  static const analogousPrimary800 = Color(0xff00831e);
+  static const analogousPrimary700 = Color(0xff00942a);
+  static const analogousPrimary600 = Color(0xff18a636);
+  static const analogousPrimary500 = Color(0xff25b53f);
+  static const analogousPrimary400 = Color(0xff4fc15d);
+  static const analogousPrimary300 = Color(0xff72cc7b);
+  static const analogousPrimary200 = Color(0xff9ddaa1);
+  static const analogousPrimary100 = Color(0xffc4e8c6);
+  static const analogousPrimary050 = Color(0xffe6f6e8);
+
+  static const analogousSecondary900 = Color(0xff004c73);
+  static const analogousSecondary800 = Color(0xff006a94);
+  static const analogousSecondary700 = Color(0xff007ba7);
+  static const analogousSecondary600 = Color(0xff0d8db9);
+  static const analogousSecondary500 = Color(0xff179ac6);
+  static const analogousSecondary400 = Color(0xff35a8cc);
+  static const analogousSecondary300 = Color(0xff54b6d3);
+  static const analogousSecondary200 = Color(0xff7fcbdf);
+  static const analogousSecondary100 = Color(0xffb0e0eb);
+  static const analogousSecondary050 = Color(0xffe0f3f7);
+}