+++ lib/view/home_screen.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:cccc/services/logger_init.dart';
 import 'package:cccc/models/user.dart';
+import 'package:cccc/view/scan_receipts_screen.dart';
 import 'package:cccc/view_models/home_screen_model.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -77,6 +78,10 @@ class HomeScreen extends ConsumerWidget {
           stretchTriggerOffset: media.size.height * 0.2,
           centerTitle: true,
           title: Text(model.today),
+          leading: IconButton(
+            icon: const Icon(Icons.photo_camera),
+            onPressed: () => ScanReceiptsScreen.show(context),
+          ),
           actions: [
             IconButton(
               onPressed: () => SettingsScreen.show(context),