+++ lib/view/transaction_detail_screen.dart
@@ -1,8 +1,14 @@
+import 'package:cccc/models/enum/payment_channel.dart';
 import 'package:cccc/models/plaid/account.dart';
 import 'package:cccc/routes/route_names.dart';
+import 'package:cccc/services/logger_init.dart';
+import 'package:cccc/theme/custom_button_theme.dart';
 import 'package:cccc/theme/text_styles.dart';
+import 'package:cccc/theme/theme_colors.dart';
+import 'package:cccc/view/scan_receipts_screen.dart';
 import 'package:cccc/view_models/transaction_detail_screen_model.dart';
 import 'package:cccc/widgets/custom_stream_builder.dart';
+import 'package:cccc/widgets/receipt_widget.dart';
 import 'package:flutter/material.dart';
 
 import 'package:cccc/models/plaid/transaction.dart';
@@ -25,7 +31,10 @@ class TransactionDetailScreen extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    logger.d('[TransactionDetail] building... ');
+
     final model = ref.watch(transactionDetailScreenModelProvider(transaction));
+    final theme = Theme.of(context);
 
     return Scaffold(
       body: CustomScrollView(
@@ -39,12 +48,12 @@ class TransactionDetailScreen extends ConsumerWidget {
               background: Stack(
                 children: [
                   Container(
-                    decoration: const BoxDecoration(
+                    decoration: BoxDecoration(
                       gradient: LinearGradient(
                         begin: Alignment.topCenter,
                         end: Alignment.bottomCenter,
                         colors: [
-                          Colors.blueAccent,
+                          theme.primaryColor,
                           Colors.black,
                         ],
                       ),
@@ -93,8 +102,27 @@ class TransactionDetailScreen extends ConsumerWidget {
           SliverToBoxAdapter(
             child: SingleChildScrollView(
               child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   const SizedBox(height: 16),
+                  ExpansionTile(
+                    initiallyExpanded: true,
+                    tilePadding: const EdgeInsets.symmetric(
+                      vertical: 8,
+                      horizontal: 16,
+                    ),
+                    textColor: theme.primaryColor,
+                    collapsedTextColor: Colors.white,
+                    leading: const Icon(Icons.receipt),
+                    childrenPadding: const EdgeInsets.symmetric(vertical: 16),
+                    title: const Text(
+                      'Items',
+                      style: TextStyles.captionNoColor,
+                    ),
+                    children: [
+                      _buildItemsWidget(context),
+                    ],
+                  ),
                   CustomStreamBuilder<Account?>(
                     stream: model.acocuntStream,
                     builder: (context, data) {
@@ -184,28 +212,30 @@ class TransactionDetailScreen extends ConsumerWidget {
                       ),
                     ],
                   ),
-                  ExpansionTile(
-                    tilePadding: const EdgeInsets.symmetric(
-                      vertical: 8,
-                      horizontal: 16,
+                  if (transaction.paymentChannel == PaymentChannel.inStore)
+                    ExpansionTile(
+                      tilePadding: const EdgeInsets.symmetric(
+                        vertical: 8,
+                        horizontal: 16,
+                      ),
+                      iconColor: Colors.white,
+                      collapsedIconColor: Colors.white,
+                      textColor: Colors.white,
+                      collapsedTextColor: Colors.white,
+                      leading: const Icon(Icons.map),
+                      title: Text(
+                        model.paymentChannel,
+                        style: TextStyles.caption,
+                      ),
+                      expandedAlignment: Alignment.centerLeft,
+                      expandedCrossAxisAlignment: CrossAxisAlignment.start,
+                      childrenPadding:
+                          const EdgeInsets.symmetric(horizontal: 24),
+                      children: [
+                        Text(model.lat, style: TextStyles.caption),
+                        Text(model.lon, style: TextStyles.caption),
+                      ],
                     ),
-                    iconColor: Colors.white,
-                    collapsedIconColor: Colors.white,
-                    textColor: Colors.white,
-                    collapsedTextColor: Colors.white,
-                    leading: const Icon(Icons.map),
-                    title: Text(
-                      model.paymentChannel,
-                      style: TextStyles.caption,
-                    ),
-                    expandedAlignment: Alignment.centerLeft,
-                    expandedCrossAxisAlignment: CrossAxisAlignment.start,
-                    childrenPadding: const EdgeInsets.symmetric(horizontal: 24),
-                    children: [
-                      Text(model.lat, style: TextStyles.caption),
-                      Text(model.lon, style: TextStyles.caption),
-                    ],
-                  ),
                   if (model.isFood)
                     const ExpansionTile(
                       tilePadding: EdgeInsets.symmetric(
@@ -227,5 +257,51 @@ class TransactionDetailScreen extends ConsumerWidget {
         ],
       ),
     );
+  }
+
+  Widget _buildItemsWidget(BuildContext context) {
+    final size = MediaQuery.of(context).size;
+
+    if (transaction.transactionItems != null) {
+      return ReceiptWidget(
+        transaction: transaction,
+        color: ThemeColors.primary500.withOpacity(0.24),
+      );
+    } else {
+      return Container(
+        height: 160,
+        width: size.width - 32,
+        decoration: BoxDecoration(
+          border: Border.all(color: Colors.white24, width: 1),
+          borderRadius: BorderRadius.circular(8),
+        ),
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              const Text(
+                'Scan your receipts now to get individual items',
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 16),
+              TextButton(
+                onPressed: () => ScanReceiptsScreen.show(context),
+                style: CustomButtonTheme.text1Primary,
+                child: Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children: const [
+                    Icon(Icons.camera_alt),
+                    SizedBox(width: 16),
+                    Text('Scan Now'),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      );
+    }
   }
 }