+++ lib/view/transactions_screen.dart
@@ -1,4 +1,5 @@
 import 'package:cccc/models/plaid/transaction.dart';
+import 'package:cccc/services/logger_init.dart';
 import 'package:cccc/widgets/transaction_list_tile.dart';
 import 'package:flutter/material.dart';
 
@@ -24,6 +25,8 @@ class TransactionsScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    logger.d('[Transactions] Screen building...');
+
     return Scaffold(
       body: CustomScrollView(
         slivers: [